feat(ingredient): add toast about the use of an ingredient in the recipe

+++ src/actions/ingredient/delete-ingredient.ts
@@ -18,12 +18,34 @@ export async function deleteIngredient(
   }
 
   try {
+    const usages = await prisma.recipeIngredient.findMany({
+      where: { ingredientId: id },
+      select: {
+        recipe: {
+          select: {
+            name: true,
+          },
+        },
+      },
+    })
+
+    if (usages.length > 0) {
+      const recipeNames = [
+        ...new Set(usages.map((usages) => usages.recipe.name)),
+      ].join(', ')
+
+      return {
+        error: `Cannot delete. This ingredient is used in the following recipes: ${recipeNames}`,
+      }
+    }
+
     const ingredient = await prisma.ingredient.delete({
       where: { id },
     })
+
     return { success: true, ingredient }
   } catch (error) {
-    console.error('Error deleting ingredient', error)
-    return { error: 'Error deleting ingredient' }
+    console.error('Error deleting ingredient:', error)
+    return { error: 'Something went wrong while deleting the ingredient' }
   }
 }
